add inception_score to metric dcgan model
