+++ example/gluon/DCgan/README.md
@@ -1,0 +1,10 @@
+# DCgan and eval
+DCgan model is from: [Unsupervised Representation Learning with Deep Convolutional Generative Adversarial Networks](https://arxiv.org/abs/1511.06434)
+
+The inception score is refer to: [openai/improved-gan
+](https://github.com/openai/improved-gan)
+
+DCgan in MXNet, and the inception_score mertic:
+
+![Generated pic](https://github.com/pengxin99/incubator-mxnet/tree/master/example/gluon/DCgan/pic/fake_img_iter_19520.png)
+![inception_socre_with_cpu](https://github.com/pengxin99/incubator-mxnet/tree/master/example/gluon/DCgan/pic/finception_score_cifar10_cpu.png)
